Kill ProjectManager only if there are 0 clients left #12
ProjectManagers got killed without any consideration for the
clients that are using the project. As a consequence
if 2 or more clients using the same project
and 1 client disconnects the whole ProjectManager got killed and the
remaining clients got errors that the project doesn't exist.

Fix this by saving the client#/project and shutdown ProjectManager
only if there is no client left.